// src/enums/advertising-review.enums.ts
export enum WalmartReviewProcessStatusEnum {
  PENDING = 'pending',
  IN_PROGRESS = 'inProgress',
  COMPLETE = 'complete',
  CANCELLED = 'cancelled',
}

export enum WalmartReviewDecisionStatusEnum {
  APPROVED = 'approved',
  REJECTED = 'rejected',
}

export enum WalmartMappedReviewStatusEnum {
  PENDING = 'Pending',
  IN_PROGRESS = 'In Progress',
  APPROVED = 'Approved',
  REJECTED = 'Rejected',
  CANCELLED = 'Cancelled',
}